Calculate costs per month

+++ src/main/java/com/meter/consumption_meter/domain/Consumption.java
@@ -1,5 +1,6 @@
 package com.meter.consumption_meter.domain;
 
+import java.math.BigInteger;
 import java.util.Date;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
@@ -12,5 +13,5 @@ import lombok.ToString;
 @EqualsAndHashCode
 public class Consumption {
     private Date timeOfReading;
-    private float numberOfKiloWattHours;
+    private BigInteger numberOfKiloWattHours;
 }
